disable host watchdog for 1010 branch
The host watchdog is dependent on an IPMI
interface which PHYP does not use. Disable
the host watchdog until the PLDM interface
is in place and Hostboot, OPAL, and PHYP are
using it.

With this change, IPMI will log the following
messages to the journal when hostboot sends down
the watchdog messages. This does not cause any
error to be logged or other issues:
rainier ipmid[332]: wd_reset: sd_bus_call: org.freedesktop.DBus.Error.FileNotFound: path or object not found

Tested:
Verified rainier booted to PHYP standby without
`obmcutil recoveryoff` being needed

Signed-off-by: Andrew Geissler <geisonator@yahoo.com>