Dual boot support: Create a BOOTKERNEL for phyp and opal
For systems that support eMMC (tacoma and rainier), create a BOOTKERNEL
to boot from either phyp or opal. Create a symlink to default to phyp.

Since both files do not fit in the NOR image, create a separate
image for emmc to be used during code update, and just keep the
phyp one on the NOR as it is today.

Change-Id: Id81fb655a31891385895c3e88107e6f33e8f13d7
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>