Revert "phosphor-ipmi-host: Remove clear-once from rainier"
This reverts commit f336cc6e37daa3d377329904de04902470dc2534.

We have transitioned to eMMC, and also the environment has been fixed
for ast2600 cards.
We need to put this package back for HW because the side effect of
removing the clear-once package is that it also removes its RDEPENDS
which is the u-boot-fw-utils package that provide the fw_setenv and
fw_printenv commands.

Change-Id: I594e35234b8d23f5ba6afcc3fdc4d4d8163a2402
Signed-off-by: Adriana Kobylak <anoo@us.ibm.com>