meta-ibm:fans:rainier: Control 2u/4u initial support
Install the fan and zone base JSON configuration files for fan control
to support the initial poweron of either the 2u or 4u from the same
image. The fan control application still uses YAML based fan and zone
configurations until the commit is merged that enables JSON support
within the fan control application. Once this commit is merged, the fan
and zone JSON configurations will be used from this commit and the YAML
config files can then be removed.

Change-Id: Ib7b891acfb8a7601d7f89d1b20eaf9013ff57c0e
Signed-off-by: Matthew Barth <msbarth@us.ibm.com>