Move aspeed2400 and aspeed2500 u-boot's to 2019.04 SDK
This is a work in progress

We're basically deleting the 2016.07 targets and replacing them
for the 2019.04 as the defaults.

Then porting the equivalent defconfigs in u-boot

This currently fails (on palmetto at least) as Yocto tries to build
u-boot with -march=armv8-a instead of -mcpu=armv5 as we would expect.

Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>