meta-aspeed: Add socsec
socsec is required for securing firmware leveraging the hardware root of
trust in the AST2600.

This recipe is set to automatic follow revisions of the default branch,
since the repository has a 'develop' branch where upstream changes are
staged before being merged into master.

Additionally, carry two patches that are needed:

0001-socsec-Ignore-non-key-data-on-PEM-key-files.patch - merged upstream
0002-ast2600-Allow-for-SPL-as-large-as-64Kb-512b.patch - submitted upstream

Change-Id: I7f44609df1d303e8211eb9286442afdfcd77c4b8
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Klaus Heinrich Kiwi <klaus@linux.vnet.ibm.com>